// graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type ImageSharpFluid = {
  base64?: Maybe<Scalars['String']>;
  tracedSVG?: Maybe<Scalars['String']>;
  aspectRatio: Scalars['Float'];
  src: Scalars['String'];
  srcSet: Scalars['String'];
  srcWebp?: Maybe<Scalars['String']>;
  srcSetWebp?: Maybe<Scalars['String']>;
  sizes: Scalars['String'];
  originalImg?: Maybe<Scalars['String']>;
  originalName?: Maybe<Scalars['String']>;
  presentationWidth: Scalars['Int'];
  presentationHeight: Scalars['Int'];
};
// ...
export type Internal = {
  content?: Maybe<Scalars['String']>;
  contentDigest: Scalars['String'];
  description?: Maybe<Scalars['String']>;
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
  ignoreType?: Maybe<Scalars['Boolean']>;
  mediaType?: Maybe<Scalars['String']>;
  owner: Scalars['String'];
  type: Scalars['String'];
};
// ...
export type JclubYaml = Node & {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  slug?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  short_name?: Maybe<Scalars['String']>;
  category?: Maybe<Scalars['String']>;
  hometown?: Maybe<Scalars['String']>;
  area?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type SitePageContext = {
  slug?: Maybe<Scalars['String']>;
  relativeDirectory?: Maybe<Scalars['String']>;
};
// ...
export type Venues = Node & {
  clubs?: Maybe<Array<Maybe<JclubYaml>>>;
  type?: Maybe<Scalars['String']>;
  slug?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  coordinates?: Maybe<Array<Maybe<Scalars['Float']>>>;
  population?: Maybe<VenuesPopulation>;
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type VenuesPopulation = {
  val_1?: Maybe<Scalars['Int']>;
  val_3?: Maybe<Scalars['Int']>;
  val_5?: Maybe<Scalars['Int']>;
  val_10?: Maybe<Scalars['Int']>;
};
// ...
export type VenueQuery = { venues?: Maybe<(
    Pick<Venues, 'name' | 'coordinates' | 'id' | 'slug' | 'type'>
    & { population?: Maybe<Pick<VenuesPopulation, 'val_1' | 'val_3' | 'val_5' | 'val_10'>>, clubs?: Maybe<Array<Maybe<Pick<JclubYaml, 'area' | 'category' | 'hometown' | 'id' | 'name' | 'short_name' | 'slug'>>>> }
  )>, allFile: { nodes: Array<{ childImageSharp?: Maybe<{ fluid?: Maybe<GatsbyImageSharpFluidFragment> }> }> } };
// ...
export type GatsbyImageSharpFluidFragment = Pick<ImageSharpFluid, 'base64' | 'aspectRatio' | 'src' | 'srcSet' | 'sizes'>;